debug: move all color registrations to debug.contribution.ts
