+++ src/vs/workbench/contrib/debug/browser/debugViewlet.ts
@@ -17,7 +17,7 @@ import { IProgressService } from 'vs/platform/progress/common/progress';
 import { IWorkspaceContextService, WorkbenchState } from 'vs/platform/workspace/common/workspace';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { IStorageService } from 'vs/platform/storage/common/storage';
-import { IThemeService, registerThemingParticipant } from 'vs/platform/theme/common/themeService';
+import { IThemeService } from 'vs/platform/theme/common/themeService';
 import { IContextMenuService, IContextViewService } from 'vs/platform/contextview/browser/contextView';
 import { IDisposable, dispose } from 'vs/base/common/lifecycle';
 import { IWorkbenchLayoutService } from 'vs/workbench/services/layout/browser/layoutService';
@@ -33,7 +33,6 @@ import { INotificationService } from 'vs/platform/notification/common/notificati
 import { IViewDescriptorService, IViewsService } from 'vs/workbench/common/views';
 import { WelcomeView } from 'vs/workbench/contrib/debug/browser/welcomeView';
 import { ToggleViewAction } from 'vs/workbench/browser/actions/layoutActions';
-import { registerColor, foreground, badgeBackground, badgeForeground, listDeemphasizedForeground, contrastBorder } from 'vs/platform/theme/common/colorRegistry';
 
 export class DebugViewPaneContainer extends ViewPaneContainer {
 
@@ -246,81 +245,3 @@ export class OpenDebugConsoleAction extends ToggleViewAction {
 		super(id, label, REPL_VIEW_ID, viewsService, viewDescriptorService, contextKeyService, layoutService, 'codicon-debug-console');
 	}
 }
-
-const debugViewExceptionLabelForeground = registerColor('debugView.exceptionLabelForeground', { dark: foreground, light: foreground, hc: foreground }, 'Foreground color for a label shown in the CALL STACK view when the debugger breaks on an exception.');
-const debugViewExceptionLabelBackground = registerColor('debugView.exceptionLabelBackground', { dark: '#6C2022', light: '#A31515', hc: '#6C2022' }, 'Background color for a label shown in the CALL STACK view when the debugger breaks on an exception.');
-const debugViewStateLabelForeground = registerColor('debugView.stateLabelForeground', { dark: foreground, light: foreground, hc: foreground }, 'Foreground color for a label in the CALL STACK view showing the current session\'s or thread\'s state.');
-const debugViewStateLabelBackground = registerColor('debugView.stateLabelBackground', { dark: '#88888844', light: '#88888844', hc: '#88888844' }, 'Background color for a label in the CALL STACK view showing the current session\'s or thread\'s state.');
-const debugViewValueChangedHighlight = registerColor('debugView.valueChangedHighlight', { dark: '#569CD6', light: '#569CD6', hc: '#569CD6' }, 'Color used to highlight value changes in the debug views (ie. in the Variables view).');
-
-registerThemingParticipant((theme, collector) => {
-	// All these colours provide a default value so they will never be undefined, hence the `!`
-	const badgeBackgroundColor = theme.getColor(badgeBackground)!;
-	const badgeForegroundColor = theme.getColor(badgeForeground)!;
-	const listDeemphasizedForegroundColor = theme.getColor(listDeemphasizedForeground)!;
-	const debugViewExceptionLabelForegroundColor = theme.getColor(debugViewExceptionLabelForeground)!;
-	const debugViewExceptionLabelBackgroundColor = theme.getColor(debugViewExceptionLabelBackground)!;
-	const debugViewStateLabelForegroundColor = theme.getColor(debugViewStateLabelForeground)!;
-	const debugViewStateLabelBackgroundColor = theme.getColor(debugViewStateLabelBackground)!;
-	const debugViewValueChangedHighlightColor = theme.getColor(debugViewValueChangedHighlight)!;
-
-	collector.addRule(`
-		/* Text colour of the call stack row's filename */
-		.debug-pane .debug-call-stack .monaco-list-row:not(.selected) .stack-frame > .file .file-name {
-			color: ${listDeemphasizedForegroundColor}
-		}
-
-		/* Line & column number "badge" for selected call stack row */
-		.debug-pane .monaco-list-row.selected .line-number {
-			background-color: ${badgeBackgroundColor};
-			color: ${badgeForegroundColor};
-		}
-
-		/* Line & column number "badge" for unselected call stack row (basically all other rows) */
-		.debug-pane .line-number {
-			background-color: ${badgeBackgroundColor.transparent(0.6)};
-			color: ${badgeForegroundColor.transparent(0.6)};
-		}
-
-		/* State "badge" displaying the active session's current state.
-		 * Only visible when there are more active debug sessions/threads running.
-		 */
-		.debug-pane .debug-call-stack .thread > .state > .label,
-		.debug-pane .debug-call-stack .session > .state > .label,
-		.debug-pane .monaco-list-row.selected .thread > .state > .label,
-		.debug-pane .monaco-list-row.selected .session > .state > .label {
-			background-color: ${debugViewStateLabelBackgroundColor};
-			color: ${debugViewStateLabelForegroundColor};
-		}
-
-		/* Info "badge" shown when the debugger pauses due to a thrown exception. */
-		.debug-pane .debug-call-stack-title > .pause-message > .label.exception {
-			background-color: ${debugViewExceptionLabelBackgroundColor};
-			color: ${debugViewExceptionLabelForegroundColor};
-		}
-
-		/* Animation of changed values in Debug viewlet */
-		@keyframes debugViewletValueChanged {
-			0%   { background-color: ${debugViewValueChangedHighlightColor.transparent(0)} }
-			5%   { background-color: ${debugViewValueChangedHighlightColor.transparent(0.9)} }
-			100% { background-color: ${debugViewValueChangedHighlightColor.transparent(0.3)} }
-		}
-
-		.debug-pane .monaco-list-row .expression .value.changed {
-			background-color: ${debugViewValueChangedHighlightColor.transparent(0.3)};
-			animation-name: debugViewletValueChanged;
-			animation-duration: 1s;
-			animation-fill-mode: forwards;
-		}
-	`);
-
-	const contrastBorderColor = theme.getColor(contrastBorder);
-
-	if (contrastBorderColor) {
-		collector.addRule(`
-		.debug-pane .line-number {
-			border: 1px solid ${contrastBorderColor};
-		}
-		`);
-	}
-});
